Fix /download-rsf/n for registers with identical consecutive entries
This is only necessary because the government-domain register
currently has consecutive entries that add duplicate items against
the same key. The indexing code assumes that this is not possible and
we therefore have to revert to using the entry table. In the future
we must either remove the duplicate entries or not reply on the
assumption that this is not possible.