Use content types without parameters in link headers
The CSV media type constant includes an encoding=UTF-8 parameter, so the whole
string isn't a valid value to pass in the `extra` argument to setLinkHeader.

If used as is, the resulting header value will quote the whole thing:
type="text/csv; charset=UTF-8".

To avoid this, just use the string "text/csv".