Re-introduce batching of records so that loads aren't slow
Refactors the batching logic in the data access layer. We used to have a
PostgresReadDataAccessLayer (for reads) and PostgresDataAcessLayer (for writes).
The former was an abstract class that had a dependency on the latter.
The idea was that whenever a read was performed in the former, the latter
wrote all it's batched (in memory) data to the database, but this was confusing.
We now have PostgresDataAccessLayer which does both reads and writes on the database
and BatchedPostgresDataAccessLayer which batches data in memory. It reads from
memory if it can, else it sends it's batched data to the database and reads from the database
using PostgresDataAcessLayer. Hopefully this is less confusing.