Add blob resource
Blob is the new name for item. Version this URL at 2 from the
beginning to indicate it's not a new thing.

Some decisions here:

- I'm currently versioning at the resource level, not the whole API level, since
  resources can evolve at different rates, and not all consumers need
  all endpoints
  - This means you can ignore breaking changes you don't care about
  - HOWEVER, changing the version of a resource requires a new version
    of any resources that link to it
  - So it may be simpler to just version the whole API anyway

- In this case, the only difference is the URL, so it's safe to link to
  blobs OR items from the original version of /entries

- Resources without a version number are implicitly version 1 of that
  resource

- We may redirect /v1/FOO to /FOO throughout, but I haven't done this
  yet

In all cases, the intent is to prevent client code from breaking
unexpectedly - where we change the API, the client should opt in to the
new version by changing the URL.