Remove indexing code (#501)
* Remove /index/* endpoints

We will no longer support these endpoints.

* Remove /index/*/download-rsf/* endpoints

We will no longer support these.

* Remove methods to query indexes from Register interface

The PostgresDataAccessLayer still operates in terms of index names
as it still uses the index table. The only indexes we still care about
are "record" and "metadata" and we translate EntryType of user or system
to these index names in the PostgresDataAccessLayer.

* Read entries and records from entry tables instead of index table

The logic for getting records and entries from the database has been
rewritten to query the entry and entry_item tables instead of the index
table (entry_system and entry_item_system for system entries and records).

* Remove indexing code

This removes all code required to store indexes in the index table.

* Migrate sha256hex column to entry table

We no longer support entries with multiple items. This is the first
step towards removing the entry_item and entry_item_system tables.

* Get entries from entry table instead of entry and entry_item tables

We no longer need to use the entry_item table because it is not possible
for an entry to contain multiple items.

* Update Entry and Record objects to contain only a single item

Ensure that we still express items as an array in the API so that
we do not create a breaking change to the API.

* Make sure we are still checking for duplicate entries being loaded

* Re-introduce batching of records so that loads aren't slow

Refactors the batching logic in the data access layer. We used to have a
PostgresReadDataAccessLayer (for reads) and PostgresDataAcessLayer (for writes).
The former was an abstract class that had a dependency on the latter.
The idea was that whenever a read was performed in the former, the latter
wrote all it's batched (in memory) data to the database, but this was confusing.
We now have PostgresDataAccessLayer which does both reads and writes on the database
and BatchedPostgresDataAccessLayer which batches data in memory. It reads from
memory if it can, else it sends it's batched data to the database and reads from the database
using PostgresDataAcessLayer. Hopefully this is less confusing.

* Rename PostgresRegister RegisterImpl

This name is still not great but is at least consistent with other
implementations of interfaces in the code base. This class has nothing
to do with Postgres so needed renaming.

* Add ADR for removing indexes

* Ensure tests pick up location for migrations

* Revert change to database rule as seems to cause failing tests on CI

* Fix tests

* add missing mock to fix test

* use different register fixture to fix redirects test