Add a command to run migrations on demand
At the moment we automatically run migrations when we start the app.
This causes problems when migrations take a long time to run,
because when doing a zero downtime deployment, cloud foundry will kill
the app if it doesn't respond to requests within a certain time.

In this case the new code is not compatable with the old database, but
the database changes are backwards compatable with the old code).

I thought it was worth creating a separate command to manually run
migrations, since this is something that is likely to happen again,
however, this creates some logistical issues, because we need to remove
the current code that runs migrations as part of the serve command, and
trigger this task elsewhere in the build, before the app is run.
I think this needs more thought, so I don't propose merging this branch.

In order to deploy master now, I'd like to deploy this branch
as a hotfix with automatic migrations disabled, run the migration
manually, and then deploy master as normal. The code on this branch
is the same as the version currently deployed, with indexes still
present.