Switch off HTTP healthcheck
As the app executes migrations on startup, the HTTP healthcheck
may fail after deploying a new version, causing the deploy to
be rolled back. This is currently blocking master from being deployed.
As a workaround, switch to a simpler healthcheck that just ensures
the process is running.

See
https://docs.cloudfoundry.org/devguide/deploy-apps/healthchecks.html#types
for a full description of health check types in cloud foundry.

This should be reverted after master is deployable again.