Use EntryView for JSON serialization when calculating proofs
The MerkleLeafStore also needs to serialize JSON, so this needs to call
the new view instead of the Entry object directly.

This is not ideal but this is a consequence of relying on JSON
serialization for our hashing. It will no longer be a problem when
we've switched to objecthash as proposed in RFC 9:
https://github.com/openregister/registers-rfcs/pull/23