Extract EntryView from Entry (#514)
* Add EntryView

This contains all the serialization/deserialization functionality
from Entry, so that the core Entry class no longer needs to handle
this. This will allow us to start versioning the view layer (i.e.
introduce multiple view classes for the same core class, without
changing the persistence layers)

* Change view factory to use EntryView instead of Entry

Change view factory to use EntryView instead of Entry with
AttributionView

* Use EntryView within Record(s)View

* Use EntryView instead of entry in list view and turtle writer

* Use EntryView when creating zip archives

The archiving code relies on JSON serialization so it should
also go via the view.

* Use EntryView for JSON serialization when calculating proofs

The MerkleLeafStore also needs to serialize JSON, so this needs to call
the new view instead of the Entry object directly.

This is not ideal but this is a consequence of relying on JSON
serialization for our hashing. It will no longer be a problem when
we've switched to objecthash as proposed in RFC 9:
https://github.com/openregister/registers-rfcs/pull/23

* Remove presentation code from Entry class

This should no longer be used for serializing JSON/CSV etc