Add new v2 endpoints
As these endpoints are a work in progress, the URLs are currently
using the "/dev" prefix, and the implementation is the same as V1.

Anyone using these endpoints should expect functionality to be
added/removed/changed. Once this version is stable, we will rename
it to /v2.

As part of this change I've moved the main resource classes (i.e. stuff
that's part of the specified API) to subpackages by version.

I've made IndexSizePagination public so it can be used in both of these
packages (previously it was package private).

As a follow up to this work, we will need to do something with the
conformance tests so that we have the ability to test multiple
versions of the API against their corresponding specification.