Introduce /blobs endpoint (CSV and JSON)
This endpoint is required by the spec but we haven't implemented it yet.
It's also available in /v1/ of the API, because we won't be making any
breaking changes to it other than the name change.

The endpoint will be ordered by item hash and will not include system
entries (i.e. it will return all unique items that are in the user
entry log).

This is an initial implementation of the endpoint without pagination
so it will need to be iterated on. The performance is also
pretty bad as it seems to do a bunch of database lookups as part of
rendering the JSON, but I'm not going to investigate this now.

This requires a new view class (ItemListView) in order to be
CSV writeable.