Remove HTML views for V2 API
It's non trivial to add multi-version support to the HTML templates (for
example, there are a lot of hyperlinks between resources that need to
take the version into account). We're not sure if we're going to
continue to support this yet, so I'm going to remove them for now and we
can implement them later if needed.