Simplify view constructors
Item and record views depend on the object they are a view of, plus
information about the valid fields for that register, which comes from
the RegisterReadOnly object.

However up until now the ItemView constructor has taken an item hash and
the Map representation of an item - `fieldValueMap` -  instead of the
actual item. This was a bit confusing because `fieldValueMap` and
`fields` are completely different things. The transformation from items
to fieldValueMap is also repeated everywhere we construct these views.

If we inject the ItemConverter into these views and just pass items
around, then we can simplify the code that calls them and all the views
will be constructed in a consistent way.

I haven't removed the original constructor, because it still used
in some tests for serialization. I'm going to leave this as they are
because
- a lot of them will go away when we drop support for non-CSV/JSON
  formats
- at the moment, the core representation of an item is coupled to
  JSON, so to create an Item object you need to do JSON manipulation (this
  will also change when we implement objecthash)