Remove unused testing class
InMemoryItemStore and InMemoryItemDAO seem to be completely
useless. They are used when testing RegisterImpl, but it can
be replaced by a stub instead. Also, we currently stub out
things that are 1 level removed from the item under test.
This moves the stubs to the direct dependencies.

I'm doing this now because I want to temporarily change the ItemStore
interface to allow for querying by old and new style hashes. This
class would also be affected by that, even though it's only used
for testing.