[WIP] Generate object hash (#540)
Implement the objecthash algorithm for blob hashes.

When loading RSF, look up the item by its v1 item hash and add the new blob hash to the item and entry.

A blob hash should ignore any null/empty strings or sets.

Also, remove the test version of EntryMapper. We can just use the real EntryMapper here. It's confusing to have two parallel ORM-type things and it means we have to update both when we change the columns.