Extract the verifiable log from the entry log
This gives the resource layer control over how a verifiable log is
instantiated, so that we can pass in different logs for different
versions of the API.

-----

Having the entry log encapsulate all the proof functionality is causing
problems now that we want to add 'version 2' proofs that use a new
algorithm for hashing entries.

It's impossible for `EntryLog` to use 2 verifiable logs (one for each
API version) without adding a parallel set of methods to the interface
(i.e. 2 methods for each type of proof). I don't want to do this because
it defeats the point of having an interface if the methods reflect
implementation details like this.

To work around this I've added `withVerfiableLog` and
`withV1VerifiableLog` methods to `RegisterContext`. `RegisterContext` is
what creates the `EntryLog`, so it has access to the same dependencies
needed to instantiate a `VerifiableLog`.

Despite this, you can only access a `VerifiableLog` within a callback,
because to construct one you need an `EntryIterator`, which doesn't have
a public constructor.  The reason for this seems to be that
`EntryIterator` talks directly to `EntryQueryDAO` and iterates over its
result sets, so you need to make sure it's closed properly after you are
finished using the `VerifiableLog`.

Once you have a `VerifiableLog`, there is no actual dependency on the
core classes for accessing data (`Register`, `EntryLog` etc) so I've
extracted the methods to a separate `ProofGenerator` class. I think
there is value in keeping the proofs code decoupled from the rest of the
application, because it's still considered an experimental feature.

This change affects the RSF parsing/generating functionality, because
previously this used to operate only on a `Register`, but now it needs
to pass around additional objects to provide proofs.