Wait ages for the application to start
See
https://docs.cloudfoundry.org/devguide/deploy-apps/manifest.html#timeout

> The timeout attribute defines the number of seconds that Cloud Foundry
> allocates for starting your application. It is related to the
> health-check-type attribute.
>
> You can increase the timeout length for very large apps that require
> more time to start. The timeout attribute defaults to 60, but you can
> set it to any value up to the Cloud Controller’s
> cc.maximum_health_check_timeout property.

I'm raising this value from the default because we force database
migrations to run on startup, and this can cause startup to be
very slow, particularly for "multi" instances, which run
migrations for multiple schemas sequentially.

This could be reduced again if we implement a better way of running
migrations.