Allow RSF command handlers to access contextual information
V1 and V2 commands are basically the same, but they are processed
differently depending on the version, so allow them to access that.

Since not all commands need the version, and not all commands need the
proof generator either, I've created a new context class to hold
these objects so it's easier to just access what's needed.