Return gone for removed formats
We removed tsv, ttl, xlsx and yaml formats at the beginning of december.
If these formats are requested via content type negotiation, ORJ
correctly returns HTTP/1.1 406 Not Acceptable. But if the user uses an
extension in the URL, such as `records.yaml`, they get a 404.

It's more appropriate to return 410 Gone, since these URLs used to be
available but were deliberately removed.