Validate entry key
https://github.com/openregister/registers-rfcs/blob/master/content/entry-key-constraints/index.md

Note that the RFC also specifies that when an item contains a field with
the same name as the register, the entry key should match this field's
value. This is currently a non-enforced convention, but I'm not going to
implement a check for it now, because I think it would require
significant refactoring of the RSF code.