Upgrade underlying OS stack
The operating system on PaaS is about to be upgraded. This means we may
find ourselves in a situation where our applications stop working.

We're bumping these ourselves before the change is enforced upon us, to
prepare ourselves for potential additional work.

The manifests are now requesting the underlying OS to be `cflinuxfs3`.